+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Changelog
 """"""""""
 
 
-v1.6.2 : CLI-clean
+v1.7.0 : CLI-clean
 ------------------
 
 **2022-05-21**
@@ -22,15 +22,18 @@ v1.6.2 : CLI-clean
     --save`` using the same file name.
   - Parameters for noise generation should new be provided under their own
     section ``[noise_opts]``; providing them under ``[simulation]`` is
-    deprecated and will be removed in v1.8.0.
+    deprecated and will be removed in v1.9.0.
 
 - Simulation:
 
   - ``'all'`` is now the same as ``'computed'`` in ``to_file`` and ``to_dict``,
     meaning the grids are stored as well.
   - Deprecation: The ``'expand'``-functionality in the gridding options is
-    deprecated and will be removed in v1.8.0. A property-complete model has to
+    deprecated and will be removed in v1.9.0. A property-complete model has to
     be provided.
+
+- Meshes: Bumped the change of the default value for ``center_on_edge`` from
+  ``True`` to ``False`` to v1.9.0, coinciding with the above deprecations.
 
 
 v1.6.1 : Max offset